Layouts are functioning again, now I need to fix the placement engine

+++ src/Router.cpp
@@ -155,7 +155,7 @@ void Router::buildRoutes() {
 	}
 	for (int type = 0; type < 2; type++) {
 		for (int i = 0; i < (int)base->stack[type].pins.size(); i++) {
-			if (base->stack[type].pins[i].outNet < base->nets.size()) {
+			if (base->stack[type].pins[i].outNet < (int)base->nets.size()) {
 				routes[base->stack[type].pins[i].outNet].addPin(base, Index(type, i));
 			} else {
 				printf("outNet is out of bounds\n");
